requirements.txt
fastapi
uvicorn[standard]
Jinja2
sqlalchemy
PyJWT
pymysql
requests